Move splash screen and redirections to Main activity
