+++ app/src/main/AndroidManifest.xml
@@ -12,15 +12,15 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Marvel_App"
         android:usesCleartextTraffic="true">
-        <activity android:name=".splash_screen.SplashActivity" android:theme="@style/SplashTheme">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
-        </activity>
         <activity
             android:name=".MainActivity"
+            android:theme="@style/SplashTheme"
             android:windowSoftInputMode="adjustPan">
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
         </activity>
     </application>
