+++ app/src/main/java/com/example/marvel_app/login/LoginFragment.kt
@@ -36,17 +36,9 @@ class LoginFragment : Fragment() {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
-        checkIfSignedIn()
         initSignUpButton()
         initSignInButton()
         initStateObserver()
-    }
-
-    private fun checkIfSignedIn() {
-        if (viewModel.checkIfUserSignedIn()) {
-            navController
-                .navigate(LoginFragmentDirections.actionLoginFragmentToHomeFragment())
-        }
     }
 
     private fun initSignUpButton() {
