+++ app/src/main/java/com/example/marvel_app/login/LoginViewModel.kt
@@ -38,11 +38,4 @@ class LoginViewModel : ViewModel() {
             }
         }
     }
-
-    fun checkIfUserSignedIn(): Boolean {
-        firebaseRepository.user?.let {
-            return true
-        }
-        return false
-    }
 }
