--- app/src/main/java/com/example/marvel_app/splash_screen/SplashActivity.kt
@@ -1,16 +1,0 @@
-package com.example.marvel_app.splash_screen
-
-import android.content.Intent
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-import com.example.marvel_app.MainActivity
-
-
-class SplashActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        val intent = Intent(this, MainActivity::class.java)
-        startActivity(intent)
-        finish()
-    }
-}
